+++ libkernel/src/fs/blk/mod.rs
@@ -1,2 +1,3 @@
 pub mod buffer;
+#[cfg(feature = "paging")]
 pub mod ramdisk;
