+++ libkernel/src/fs/filesystems/mod.rs
@@ -1,3 +1,4 @@
 pub mod ext4;
 pub mod fat32;
+#[cfg(feature = "alloc")]
 pub mod tmpfs;
