+++ libkernel/src/fs/filesystems/tmpfs.rs
@@ -11,7 +11,7 @@ use crate::{
         PAGE_SIZE,
         address::{AddressTranslator, VA},
         allocators::phys::PageAllocGetter,
-        page::ClaimedPage,
+        claimed_page::ClaimedPage,
     },
     sync::spinlock::SpinLockIrq,
 };
