+++ libkernel/src/memory/mod.rs
@@ -1,9 +1,18 @@
 pub mod address;
+#[cfg(feature = "paging")]
+pub mod address_space;
+#[cfg(feature = "alloc")]
 pub mod allocators;
+#[cfg(feature = "alloc")]
+pub mod claimed_page;
+#[cfg(feature = "kbuf")]
 pub mod kbuf;
 pub mod page;
+#[cfg(feature = "paging")]
 pub mod permissions;
+#[cfg(feature = "paging")]
 pub mod pg_offset;
+#[cfg(feature = "proc_vm")]
 pub mod proc_vm;
 pub mod region;
